+++ src/repository/LocalRepositories.ts
@@ -1,5 +1,6 @@
 import simpleGit from "simple-git";
 import LocalRepositoryList from "./LocalRepositoryList";
+import { pullRepositoryIfNewer } from ".";
 
 /**
  * Local repositories
@@ -54,24 +55,9 @@ export default class LocalRepositories {
 	 */
 	async pullIfNewer() {
 		for (const repository of this.repositoryList.repositories) {
-			const git = simpleGit(repository.path);
-
-			// Fetch the latest metadata
-			await git.fetch();
-
-			// Get the latest local commit date
-			const localLog = await git.log({ n: 1 });
-			const localDate = localLog.latest
-				? new Date(localLog.latest.date)
-				: new Date(0);
-
-			// Get the latest remote commit date (tracking branch)
-			// We assume 'origin/main' or 'origin/master'.
-			// Using '@{u}' (upstream) is the most robust way to reference the remote tracking branch.
-			const remoteLog = await git.log(["-1", "@{u}"]);
-			const remoteDate = remoteLog
-				? new Date(remoteLog.latest.date)
-				: new Date(0);
+			try {
+				await pullRepositoryIfNewer(repository.path);
+			} catch (err) {}
 		}
 	}